+++ src/main/java/ru/volnenko/cloud/testhub/model/Jacoco.java
@@ -34,7 +34,7 @@ public class Jacoco {
 
     @NonNull
     @Column(name = "coverage", nullable = false)
-    private Float coverage = 0F;
+    private double coverage = 0;
 
     @CreationTimestamp
     @Temporal(TemporalType.TIMESTAMP)
